changed long to int

+++ src/StringCalculator.java
@@ -2,25 +2,25 @@ import java.util.ArrayList;
 import java.util.Arrays;
 
 public class StringCalculator {
-    public static long add(String numbers){
+    public static int add(String numbers){
         // Check if the input string is empty
         if(numbers.isEmpty()){
             return 0;
         }
 
         //create a list to store all the integers
-        ArrayList<Long> list = new ArrayList<>();
+        ArrayList<Integer> list = new ArrayList<>();
 
         //split the string by commas, convert the split string to integers and add them to the list
         String[] numberStrings = numbers.split(",");
         for (String numberString : numberStrings) {
-            list.add(Long.parseLong(numberString));
+            list.add(Integer.parseInt(numberString));
         }
 
         //calculate the sum
-        long sum = 0;
+        int sum = 0;
 
-        for (long number : list) {
+        for (int number : list) {
             sum += number;
         }
